@@ -1,5 +1,6 @@
 {
   "name": "redux-remodel",
+  "private": true,
   "version": "0.1.1",
   "description": "Create reducers from simple declarative data models",
   "keywords": [
@@ -13,57 +14,50 @@
   "author": "gingerich",
   "license": "MIT",
   "repository": "gingerich/redux-remodel",
-  "main": "dist/index.js",
-  "module": "dist/index.es.js",
-  "jsnext:main": "dist/index.es.js",
   "engines": {
-    "node": ">=8",
+    "node": ">=10",
     "npm": ">=5"
   },
   "scripts": {
-    "test": "cross-env CI=1 react-scripts test --env=jsdom",
-    "test:watch": "react-scripts test --env=jsdom",
-    "build": "rollup -c",
-    "start": "rollup -c -w",
-    "prepare": "npm run build",
-    "predeploy": "cd example && npm install && npm run build",
-    "deploy": "gh-pages -d example/build"
-  },
-  "peerDependencies": {
-    "prop-types": "^15.5.4",
-    "react": "^15.0.0 || ^16.0.0",
-    "react-dom": "^15.0.0 || ^16.0.0"
+    "test": "jest",
+    "test:ci": "jest --ci",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "lint": "eslint --cache --format codeframe --ext mjs,jsx,js src",
+    "lint:fix": "npm run lint -- --fix",
+    "prettier:check": "prettier --ignore-path .eslintignore --check \"**/*.{css,html,js,jsx,json,md,yaml,yml}\"",
+    "prettier:fix": "prettier --ignore-path .eslintignore --write \"**/*.{css,html,js,jsx,json,md,yaml,yml}\"",
+    "build": "lerna run build"
   },
   "devDependencies": {
-    "@svgr/rollup": "^2.4.1",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.5",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "cross-env": "^5.1.4",
-    "eslint": "^5.0.1",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-config-standard-react": "^6.0.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-node": "^7.0.1",
-    "eslint-plugin-promise": "^4.0.0",
-    "eslint-plugin-react": "^7.10.0",
-    "eslint-plugin-react-hooks": "^2.3.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "gh-pages": "^1.2.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-scripts": "^1.1.4",
-    "rollup": "^0.64.1",
-    "rollup-plugin-babel": "^3.0.7",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-generate-package-json": "^3.2.0",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-postcss": "^1.6.2",
-    "rollup-plugin-url": "^1.4.0"
+    "@neutrinojs/eslint": "^9.1.0",
+    "@neutrinojs/jest": "^9.1.0",
+    "@neutrinojs/library": "^9.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-plugin-prettier": "^3.1.2",
+    "jest": "^24.0.0",
+    "lerna": "^3.20.2",
+    "neutrino": "^9.1.0",
+    "prettier": "^1.19.1",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.11"
+  },
+  "resolutions": {
+    "webpack": "4.41.2",
+    "react-dev-utils": "10.1.0"
+  },
+  "workspaces": {
+    "development": true,
+    "packages": [
+      "packages/*",
+      "example"
+    ],
+    "nohoist": [
+      "packages/**/webpack",
+      "**/babel-loader",
+      "**/babel-jest"
+    ]
   },
   "files": [
     "dist"
@@ -71,11 +65,5 @@
   "exports": {
     "./": "./dist/",
     ".": "./dist"
-  },
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.0.4",
-    "immer": "^5.0.0",
-    "reduce-reducers": "^1.0.4",
-    "redux": "^4.0.4"
   }
 }
