@@ -1,81 +1,70 @@
-import React from "react";
-import { TodoItem } from "./TodoItem";
-import { TodoFooter } from "./TodoFooter";
-import { TodoInput } from "./TodoInput";
-import Store, { useFetchUserTodos, useSetAppView } from "./store";
-import { ALL_TODOS } from "./constants";
+import React, { useState } from 'react';
+import { TodoItem } from './TodoItem';
+import { TodoFooter } from './TodoFooter';
+import { TodoInput } from './TodoInput';
+import { Store, useSetAppView } from './store';
+import { ALL_TODOS } from './constants';
 
-import "todomvc-app-css/index.css";
+import 'todomvc-app-css/index.css';
 
-export const App = ({ userId, view = ALL_TODOS }) => {
-  useSetAppView(view)
-  useFetchUserTodos(userId);
-  return <TodoApp />;
+export const App = ({ view = ALL_TODOS }) => {
+	useSetAppView(view);
+	return <TodoApp />;
 };
 
-const mapStoreToProps = ([state, dispatch]) => ({
-  ...state,
-  ...dispatch // spread actions into props
-});
+export const TodoApp = () => {
+	const [ state, setState ] = useState({ editing: null });
+	const [ { active, visibleTodos, view }, dispatch ] = Store.useStore();
 
-export const TodoApp = Store.withStore(mapStoreToProps)(
-  class TodoApp extends React.Component {
-    state = {
-      editing: null
-    };
+	const toggleAll = (event) => {
+		const { checked } = event.target;
+		dispatch.toggleAll(checked);
+	};
 
-    toggleAll = event => {
-      const checked = event.target.checked;
-      this.props.toggleAll(checked);
-    };
+	const edit = (todo) => {
+		setState({ editing: todo.id });
+	};
 
-    edit = todo => {
-      this.setState({ editing: todo.id });
-    };
+	const save = (todo, newTitle) => {
+		dispatch.save(todo, newTitle);
+		setState({ editing: null });
+	};
 
-    save = (todo, newTitle) => {
-      this.props.save({ ...todo, title: newTitle });
-      this.setState({ editing: null });
-    };
+	const cancel = () => {
+		setState({ editing: null });
+	};
 
-    cancel = () => {
-      this.setState({ editing: null });
-    };
+	console.log('hghjk', view, visibleTodos);
 
-    render() {
-      const { visibleTodos, active, addTodo } = this.props;
+	const todoItems = visibleTodos.map((todo) => (
+		<TodoItem
+			key={todo.id}
+			todo={todo}
+			onEdit={() => edit(todo)}
+			editing={state.editing === todo.id}
+			onSave={(newTitle) => save(todo, newTitle)}
+			onCancel={cancel}
+		/>
+	));
 
-      const todoItems = visibleTodos.map(todo => (
-        <TodoItem
-          key={todo.id}
-          todo={todo}
-          onEdit={() => this.edit(todo)}
-          editing={this.state.editing === todo.id}
-          onSave={newTitle => this.save(todo, newTitle)}
-          onCancel={this.cancel}
-        />
-      ));
-
-      return (
-        <div>
-          <header className="header">
-            <h1>Todos</h1>
-            <TodoInput onSave={addTodo} />
-          </header>
-          <section className="main">
-            <input
-              id="toggle-all"
-              className="toggle-all"
-              type="checkbox"
-              onChange={this.props.toggleAll}
-              checked={active.length === 0}
-            />
-            <label htmlFor="toggle-all" />
-            <ul className="todo-list">{todoItems}</ul>
-          </section>
-          <TodoFooter nowShowing={this.props.view} />
-        </div>
-      );
-    }
-  }
-);
+	return (
+		<div>
+			<header className="header">
+				<h1>todos</h1>
+				<TodoInput onSave={dispatch.addTodo} />
+			</header>
+			<section className="main">
+				<input
+					id="toggle-all"
+					className="toggle-all"
+					type="checkbox"
+					onChange={toggleAll}
+					checked={active.length === 0}
+				/>
+				<label htmlFor="toggle-all" />
+				<ul className="todo-list">{todoItems}</ul>
+			</section>
+			<TodoFooter nowShowing={view} />
+		</div>
+	);
+};
