@@ -1,6 +1,6 @@
-export const ALL_TODOS = "all";
-export const ACTIVE_TODOS = "active";
-export const COMPLETED_TODOS = "completed";
+export const ALL_TODOS = 'all';
+export const ACTIVE_TODOS = 'active';
+export const COMPLETED_TODOS = 'completed';
 
 export const ENTER_KEY = 13;
 export const ESCAPE_KEY = 27;
