@@ -1,6 +1,6 @@
-import { createModel, select } from 'redux-remodel'
-import todos from './todos'
-import { ALL_TODOS, ACTIVE_TODOS, COMPLETED_TODOS } from '../constants'
+import { createModel } from 'redux-remodel';
+import todos from './todos';
+import { ALL_TODOS, ACTIVE_TODOS, COMPLETED_TODOS } from '../constants';
 
 // Compose reducer functions wih slices
 export const app = {
@@ -8,26 +8,26 @@ export const app = {
     todos
   },
   computed: {
-    active: select(todos => todos.filter(t => !t.completed), 'todos'),
-    completed: select(todos => todos.filter(t => t.completed), 'todos'),
-    visibleTodos: (state) => {
+    active: ({ todos }) => todos.filter(t => !t.completed),
+    completed: ({ todos }) => todos.filter(t => t.completed),
+    visibleTodos: state => {
       switch (state.view) {
         case ACTIVE_TODOS:
-            return state.active
+          return state.active;
         case COMPLETED_TODOS:
-            return state.completed
+          return state.completed;
         default:
-            return state.todos
+          return state.todos;
       }
     }
   },
   actions: {
     setView: (state, { payload: view }) => {
-      state.view = view
+      state.view = view;
     }
   }
-}
+};
 
-export const initialState = { view: ALL_TODOS }
+export const initialState = { view: ALL_TODOS };
 
-export default createModel(initialState, app)
+export default createModel(initialState, app);
