@@ -1,7 +1,7 @@
-import { createModel } from '.'
+import { createModel } from '..';
 
 describe('createModel', () => {
   it('should require initialState to be defined', () => {
-    expect(() => createModel()).toThrow()
-  })
-})
+    expect(() => createModel()).toThrow();
+  });
+});
